% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{cm_loglik}
\alias{cm_loglik}
\title{Log-likelihood of data under a cognitive model}
\usage{
cm_loglik(model, data, ...)
}
\arguments{
\item{model}{A \code{cogmodel} object.}

\item{data}{A data frame in the model's expected format.}

\item{...}{Passed to methods.}
}
\description{
Log-likelihood of data under a cognitive model
}
