% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{cm_simulate}
\alias{cm_simulate}
\title{Simulate data from a cognitive model}
\usage{
cm_simulate(model, ...)
}
\arguments{
\item{model}{A \code{cogmodel} object.}

\item{...}{Passed to methods.}
}
\description{
Simulate data from a cognitive model
}
