% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogzoo-package.R
\docType{package}
\name{cogzoo-package}
\alias{cogzoo}
\alias{cogzoo-package}
\title{cogzoo: A Community Zoo of Computational Cognitive Models}
\description{
Every model in the zoo is an object of class \code{cogmodel} (plus a subclass
for the specific model, e.g. \code{gcm}) and implements a subset of the
generics \code{\link[=cm_simulate]{cm_simulate()}}, \code{\link[=cm_loglik]{cm_loglik()}}, \code{\link[=cm_fit]{cm_fit()}}, and \code{\link[=cm_predict]{cm_predict()}}.
Datasets ship under \verb{inst/extdata/<dataset_id>/} alongside a YAML sidecar
validated by \code{\link[=validate_cogdata]{validate_cogdata()}}. See \code{vignette("contributing")} for how
to add a new model or dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gkacherg/cogzoo}
  \item Report bugs at \url{https://github.com/gkacherg/cogzoo/issues}
}

}
\author{
\strong{Maintainer}: George Kachergis \email{george.kachergis@gmail.com}

Authors:
\itemize{
  \item George Kachergis \email{george.kachergis@gmail.com}
}

}
\keyword{internal}
