% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-registry.R
\name{list_models}
\alias{list_models}
\title{List registered models}
\usage{
list_models(domain = NULL)
}
\arguments{
\item{domain}{Optional domain to filter by (see \code{\link[=register_model]{register_model()}}).}
}
\value{
A named list of model registry entries.
}
\description{
List registered models
}
