% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-registry.R
\name{model_info}
\alias{model_info}
\title{Look up a registered model's metadata}
\usage{
model_info(name)
}
\arguments{
\item{name}{Model identifier, as passed to \code{\link[=register_model]{register_model()}}.}
}
\description{
Look up a registered model's metadata
}
